Create Adventurer's Treasure.c
An adventurer returning from a long journey with a bag of treasure. The adventurer wants to sort the treasure to see what they have, so they decide to use quick sort to sort the items in their bag. The program then performs quick sort on an array of integers representing the treasure and prints out the sorted array. Finally, the adventurer is pleased with the sorted treasure and can now count their wealth.

Sample Input:
5 		        // Length of the array treasure
10 5 40 2 40 	// Treasure


Sample Output:
After sorting:
2 5 10 40 40 

The adventurer is pleased with the sorted treasure and can now count their wealth.