Create Greedy Technique-Thief Problem.c
There is a thief who has to steal valuable items from a house but can only carry a limited weight of items. The thief wants to maximize the total value of the items he steals while staying under the weight limit. Write a program that uses the greedy approach to solve the problem. 
Hint: The program then uses the qsort() function to sort the items by their value-to-weight ratio in descending order. This is where the greedy technique comes in - by sorting the items in this way, the thief will always take the most valuable item that fits in his bag first.
The program then loops through the sorted items and adds them to the thief's bag until the weight limit is reached. If an item is too heavy to be added, it is skipped. The program prints a message for each item that is taken.
Note: Note that this program assumes that the user enters valid input (e.g. positive weights and values, weight limit greater than or equal to the weight of the heaviest item).

Sample Input:
5 	//Number of items in the house
1 10	//Input of weights and values of the items
2 20
3 20
4 30
5 40
6	//Total weight the thief can carry



Sample Output:
50