Create Greedy Truck Problem.c
 A truck driver needs to transport goods from one city to another. The truck has a maximum weight capacity of W. There are n items to be transported, each with a weight w[i] and a corresponding value v[i]. The driver wants to maximize the total value of the transported items while staying within the weight capacity of the truck. 
In this program, the fractionalKnapsack function takes an array w of item weights, an array v of item values, and the weight capacity W of the truck, and returns the maximum possible value of items that can be transported within the weight capacity W. The main function is an example driver program that creates an example input, calls fractionalKnapsack to compute the maximum possible value, and prints the result. Note that the fractionalKnapsack function assumes that the weights and values are given in the same order, i.e., w[i] corresponds to v[i] for all i. If the input arrays are not in the same order, the function will need to be modified accordingly.

Note: The truck can take fractions of each item if not the whole.


Sample Input:
5 	//Number of items 
1 10	//Input of weights and values of the items
2 20
3 20
4 30
5 40
10	//Total weight 


Sample Output:
85.00