Create Coin Change Problem.c
Write a C  program that uses dynamic programming to find the minimum number of coins required to make a given amount of money. The program takes the amount as input and returns the minimum number of coins required to make the given amount where the coins can occur in either 1, 5, 10, 25 Rupees. The program uses an array to store the minimum number of coins required to make each possible amount from 0 to the given amount. The program then iterates through the array, checking each coin and updating the array if a smaller number of coins is found. Finally, the program returns the minimum number of coins required to make the given amount.

Sample Input:
36
Sample Output:
3